pyproject.toml
[project]
name = "yt-playlist-sum-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "cachetools>=6.2.2",
    "email-validator>=2.3.0",
    "fastapi>=0.124.0",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "google-generativeai>=0.8.5",
    "greenlet>=3.3.0",
    "groq>=0.37.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.12.8",
    # RAG dependencies
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "pgvector>=0.4.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
